+++ packages/eyes-selenium/test/coverage/custom/TestEyesDifferentRunners.spec.js
@@ -12,7 +12,6 @@ const {assertImages, getApiData} = require('./util/ApiAssertions')
 const assert = require('assert')
 const batch = getBatch()
 
-
 describe('TestEyesDifferentRunners', () => {
   let webDriver, eyes