+++ packages/eyes-webdriverio-4/CHANGELOG.md
@@ -7,6 +7,7 @@
 - updated to @applitools/dom-utils@4.7.16
 - updated to @applitools/eyes-sdk-core@10.1.1
 - updated to @applitools/visual-grid-client@13.8.0
+- handle mobile devices during working with frames
 
 ## 2.14.0 - 2020/4/30